instance/dockers/ocr_model/pyproject.toml
[project]
name = "ocr-model"
version = "0.1.0"
description = "OCR model container using EasyOCR for SwarmPilot Instance Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
